requirements.txt
pytest
requests
selenium
locust
pytest-axe
pact-python
pytest-html